requirements.txt
autopep8==1.5.7
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
PyYAML==5.4.1
toml==0.10.2
websockets==9.1
Werkzeug==2.0.1
